XSD and Schematron Vaidation - create new validation class - fix bug in existing Schematron class.
